wasm/webapi/esm-integration/module-parse-error.tentative.html

<!DOCTYPE html>
<title>Handling of importing invalid WebAssembly module</title>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    setup({allow_uncaught_exception: true});

    window.log = [];

    window.addEventListener("error", ev => log.push(ev.error));

    const test_load = async_test(
        "Test that import of invalid WebAssembly module leads to WebAssembly.CompileError on window.");
    window.addEventListener("load", test_load.step_func_done(ev => {
      assert_equals(log.length, 1);
      assert_equals(log[0].constructor, WebAssembly.CompileError);
    }));

    function unreachable() { log.push("unexpected"); }
</script>
<script type="module" src="./resources/invalid-module.wasm" onerror="unreachable()"></script>
