third_party/blink/web_tests/external/wpt/wasm/webapi/esm-integration/module-parse-error.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Test that imports of invalid WebAssembly modules leads to WebAssembly.CompileError on window.
  assert_equals: expected function "function CompileError() { [native code] }" but got function "function String() { [native code] }"
Harness: the test ran to completion.
